feature(sign up): setup test structure and write basic tests for sign up form and text input component
