+++ package.json
@@ -4,7 +4,8 @@
   "description": "A simple document management system api, with a web app which consumes the api",
   "main": "index.js",
   "scripts": {
-    "mocha": "mocha --timeout 5000 --compilers js:babel-core/register server/tests/**/*.spec.js client/tests",
+    "mocha": "mocha --timeout 5000 --compilers js:babel-core/register server/tests/**/*.spec.js client/tests/**/*.spec.jsx",
+    "mocha:client": "mocha --timeout 5000 --compilers js:babel-core/register client/tests/**/*.spec.jsx",
     "migrate": "sequelize db:migrate",
     "seed": "sequelize db:seed:all",
     "test": "NODE_ENV=test npm run mocha",
@@ -16,7 +17,7 @@
     "start": "node dist/app.js",
     "develop:server": "NODE_ENV=development npm-run-all --parallel lint:watch open:server",
     "open:server": "nodemon server/app.js --exec babel-node",
-    "cover": "rm -rf coverage/ && NODE_ENV=test babel-node ./node_modules/babel-istanbul/lib/cli.js cover _mocha -- -R spec server/tests/**/*.spec.js client/tests",
+    "cover": "rm -rf coverage/ && NODE_ENV=test babel-node ./node_modules/babel-istanbul/lib/cli.js cover _mocha -- -R spec server/tests/**/*.spec.js client/tests/**/*.spec.jsx",
     "coveralls": "npm run cover && cat ./coverage/lcov.info | coveralls"
   },
   "repository": {
@@ -42,7 +43,6 @@
     "coveralls": "^2.11.15",
     "cross-env": "1.0.7",
     "css-loader": "0.23.1",
-    "enzyme": "2.2.0",
     "eslint": "^3.14.1",
     "eslint-config-airbnb": "^14.0.0",
     "eslint-plugin-import": "^2.2.0",
@@ -50,8 +50,6 @@
     "eslint-plugin-react": "^6.9.0",
     "eslint-watch": "2.1.11",
     "eventsource-polyfill": "0.9.6",
-    "expect": "1.19.0",
-    "express": "4.13.4",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.8.5",
     "gulp": "^3.9.1",
@@ -65,7 +63,7 @@
     "mocha-lcov-reporter": "^1.2.0",
     "nock": "8.0.0",
     "npm-run-all": "1.8.0",
-    "react-addons-test-utils": "15.0.2",
+    "react-addons-test-utils": "^15.4.2",
     "redux-immutable-state-invariant": "1.2.3",
     "redux-mock-store": "1.0.2",
     "rimraf": "2.5.2",
@@ -92,7 +90,9 @@
     "bootstrap": "3.3.6",
     "chai": "^3.5.0",
     "dotenv": "^4.0.0",
+    "expect": "1.19.0",
     "express": "^4.14.1",
+    "enzyme": "2.2.0",
     "faker": "^3.1.0",
     "fm-log": "^4.4.3",
     "jquery": "2.2.3",
